+++ src/hooks/useSocketAuth.server.ts
@@ -17,16 +17,16 @@ export async function useSocketAuth(cookie: string): Promise<string> {
 
   try {
     const vToken = jwt.verify(cookie, secret.toString()) as User;
-    const user = await processQuery<User>(
+    const [user] = await processQuery<User>(
       "SELECT `id`, `username`, `whitelist_status` FROM `users` WHERE `id` = ?",
       [vToken.id],
     );
 
-    if (!user[0]) {
+    if (!user) {
       return Promise.reject({ error: "user was not found", status: "error" });
     }
 
-    if (user[0].whitelist_status === Whitelist.Pending) {
+    if (user.whitelist_status === Whitelist.Pending) {
       return Promise.reject({
         error: "user is still pending access for CAD",
         status: "error",